requirements.txt
openai-whisper
sounddevice
numpy
fastmcp
mcp
soundfile
